2023/05/02 后端-修改team实体类属性的类型

+++ dcims-service/src/main/java/com/ruoyi/system/domain/DcimsTeam.java
@@ -88,7 +88,7 @@ public class DcimsTeam extends BaseEntity {
     /**
      * 下一级审核人工号
      */
-    private Integer nextAuditId;
+    private Long nextAuditId;
     /**
      * 版本
      */
